Revert "Do a big ol' banner every morning"
This was a half-cut commit. What was I even thinking? It's shit.

It's a total hack, too- it doesn't even show up in the commands list-
I even had to hack around displaying commands because it was so ill
thought out.

Instead, I've created a noddy little bot that prints an MOTD. Same
effect, and without compromising this project with a hidden command.

This reverts commit 29a3bfcbecc99a2ad3137bc6cf1d0ef079c26f72.